split components into 2
