+++ components/LeaguePage/ClassView.tsx
@@ -1,26 +1,15 @@
-import React, { useContext, useState } from "react";
-import { Class } from "../../models/class";
-import { ClassCard } from "./ClassCard";
+import React, {useState} from "react";
+import {Class} from "../../models/class";
 import styles from "./LeagueViews.module.css";
-import useSWR from "swr";
-import { APIContext } from "../../context/APIContext";
-import { Empty } from "../util/Empty";
-import { Loader } from "../util/Loader";
-import { Error } from "../util/Error";
+import {ClassScroll} from "./ClassScroll";
 
 type ClassViewProp = {
   classes: Class[];
 };
 
-type OrderBy = {
+export type OrderBy = {
   alpha: boolean;
   level: boolean;
-};
-
-type ClassScrollProp = {
-  searchQuery: string;
-  orderBy: OrderBy;
-  selectedLevels: Set<number>;
 };
 
 const filters = [
@@ -34,84 +23,6 @@ const filters = [
   "Level 7",
   "Level 8",
 ];
-
-const ClassScroll: React.FC<ClassScrollProp> = ({ searchQuery, orderBy, selectedLevels }) => {
-  const client = useContext(APIContext);
-  const { data: classes, error } = useSWR("/api/class", () => client.getAllClasses());
-
-  if (error) return <Error />;
-  if (!classes) return <Loader />;
-
-  const checkIfLevelSelected = (selectedLevels: Set<number>, tempClass: Class): boolean => {
-    if (selectedLevels.size === 0) {
-      return true;
-    }
-
-    let selected = false;
-
-    // construct array of numbers from minLevel to maxLevel (inclusive) and check if one of these levels is a
-    // level the user wants
-    Array(tempClass.maxLevel - tempClass.minLevel + 1)
-      .fill(0)
-      .map((_, i) => i)
-      .map((val) => val + tempClass.minLevel)
-      .forEach((level) => {
-        if (selectedLevels.has(level)) selected = true;
-      });
-
-    return selected;
-  };
-
-  const sortBy = (a: Class, b: Class): number => {
-    if (orderBy.alpha) {
-      if (a.name < b.name) {
-        return -1;
-      }
-      if (a.name > b.name) {
-        return 1;
-      } else {
-        return 0;
-      }
-    } else if (orderBy.level) {
-      if (a.minLevel < b.minLevel) {
-        return -1;
-      }
-      if (a.minLevel > b.minLevel) {
-        return 1;
-      } else {
-        return 0;
-      }
-    }
-    return 0;
-  };
-
-  const filteredClasses = classes
-    .filter(
-      (currClass) =>
-        currClass.name.includes(searchQuery) && checkIfLevelSelected(selectedLevels, currClass)
-    )
-    .sort((a, b) => {
-      return sortBy(a, b);
-    });
-
-  if (classes.length == 0) return <Empty userType="Classes" />;
-
-  return (
-    <>
-      {filteredClasses.map((currClass: Class) => (
-        <ClassCard
-          key={currClass.eventInformationId}
-          name={currClass.name}
-          minLevel={currClass.minLevel}
-          maxLevel={currClass.maxLevel}
-          rrstring={currClass.rrstring}
-          startTime={currClass.startTime}
-          endTime={currClass.endTime}
-        />
-      ))}
-    </>
-  );
-};
 
 const ClassView: React.FC<ClassViewProp> = () => {
   const [searchBox, setSearchBox] = useState("");
